minimise runtime when re-running

+++ src/run_pipeline.sh
@@ -143,16 +143,18 @@ echo -e "    STEP 1 completed at $s1_time\n"
 # Step 2: taxon ID/protein screening
 echo " >> STEP 2: Checking regulated pathogen proteins..."
 if [ "$BLAST" = 1 ]; then
-    echo -e "\t...running run_blastx.sh"
-    ${CM_DIR}/run_blastx.sh -d $DB_PATH/nr_blast/nr -q $QUERY -o ${OUTPUT}.nr -t $THREADS
-    echo -e "\t...checking blast results"
-    python ${CM_DIR}/check_reg_path.py -i ${OUTPUT}.nr.blastx --benign-db $DB_PATH/benign_db/ --biorisk-db $DB_PATH/biorisk_db/
+    if [ ! -f "${OUTPUT}".nr.blastx ]; then
+        echo -e "\t...running run_blastx.sh"
+        ${CM_DIR}/run_blastx.sh -d $DB_PATH/nr_blast/nr -q $QUERY -o ${OUTPUT}.nr -t $THREADS
+        echo -e "\t...checking blast results"
+        python ${CM_DIR}/check_reg_path.py -i ${OUTPUT}.nr.blastx --benign-db $DB_PATH/benign_db/ --biorisk-db $DB_PATH/biorisk_db/
 else 
-    echo -e "\t...running run_diamond.sh"
-    ${CM_DIR}/run_diamond.sh -d $DB_PATH/nr_dmnd/ -i $QUERY -o ${OUTPUT}.nr -t $THREADS -p $PROCESSES 
-    cat ${OUTPUT}.nr* > ${OUTPUT}.nr.dmnd
-    echo -e "\t...checking diamond results"
-    python ${CM_DIR}/check_reg_path.py -i ${OUTPUT}.nr.dmnd --benign-db $DB_PATH/benign_db/ --biorisk-db $DB_PATH/biorisk_db/
+    if [ ! -f "${OUTPUT}".nr.dmnd ]; then
+       echo -e "\t...running run_diamond.sh"
+        ${CM_DIR}/run_diamond.sh -d $DB_PATH/nr_dmnd/ -i $QUERY -o ${OUTPUT}.nr -t $THREADS -p $PROCESSES 
+        cat ${OUTPUT}.nr* > ${OUTPUT}.nr.dmnd
+        echo -e "\t...checking diamond results"
+        python ${CM_DIR}/check_reg_path.py -i ${OUTPUT}.nr.dmnd --benign-db $DB_PATH/benign_db/ --biorisk-db $DB_PATH/biorisk_db/
 fi
 
 s2_time=$(date)
